pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic-search-rag"
version = "0.2.0"
description = "Production-style semantic search and retrieval augmented generation service."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.34.0",
  "sentence-transformers==3.3.1",
  "numpy==2.2.1",
  "pydantic==2.10.4",
  "faiss-cpu==1.9.0.post1",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.4",
]

[project.scripts]
rag-ingest = "rag_search.ingest:main"
rag-search = "rag_search.search:main"
rag-evaluate = "rag_search.evaluate:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
